refactor: extract production labour tier-row widgets into part (#3878)
Split per-tier row chrome from production_labour_section.dart into a
library part so the host file stays under the repo size target without
changing Labour Controls behavior.

Refs #3878
