+++ app/lib/features/game/widgets/production/production_labour_section.dart
@@ -2,6 +2,8 @@
 // immediate disband). SPEC/ui/production-panel.md § Labour Controls,
 // SPEC/game/workers-and-population.md.
 
+library;
+
 import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:flutter/material.dart';
 
@@ -12,8 +14,7 @@ import '../chrome/ct_danger_text_button.dart';
 import 'production_allocation_row_buttons.dart';
 import 'production_labour_helpers.dart';
 
-const _uiIconLabourIncrement = 'ui_icon_production_alloc_increment.png';
-const _uiIconLabourDecrement = 'ui_icon_production_alloc_decrement.png';
+part 'production_labour_section_tier_row.dart';
 
 /// Labour controls section appended to the Workers section of the
 /// Available subpanel. Renders one row per worker tier with recruit/train
@@ -58,265 +59,6 @@ class ProductionLabourSection extends StatelessWidget {
             ),
           ),
       ],
-    );
-  }
-}
-
-class _ProductionLabourTierRow extends StatelessWidget {
-  const _ProductionLabourTierRow({
-    required this.data,
-    required this.callbacks,
-    required this.canEdit,
-    required this.l10n,
-    required this.theme,
-  });
-
-  final ProductionLabourTierRowData data;
-  final ProductionLabourCallbacks callbacks;
-  final bool canEdit;
-  final AppLocalizations l10n;
-  final ThemeData theme;
-
-  String _tierName() {
-    switch (data.tier) {
-      case WorkerTier.peasant:
-        return l10n.production_workers_peasants;
-      case WorkerTier.apprentice:
-        return l10n.production_workers_apprentices;
-      case WorkerTier.journeyman:
-        return l10n.production_workers_journeymen;
-      case WorkerTier.master:
-        return l10n.production_workers_masters;
-    }
-  }
-
-  String _tierLabelWithUnlockState() {
-    final tierName = _tierName();
-    final state = data.techUnlocked
-        ? l10n.production_labourTierUnlocked
-        : l10n.production_labourTierLocked;
-    return l10n.production_labourTierLabel(tierName, state);
-  }
-
-  String _appendTooltip() {
-    final label = _tierName();
-    return data.tier == WorkerTier.peasant
-        ? l10n.production_labourRecruitTier(label)
-        : l10n.production_labourTrainTier(label);
-  }
-
-  Widget _buildQueuedSegment() {
-    if (data.queuedCount <= 0) {
-      return const SizedBox.shrink();
-    }
-    return Padding(
-      padding: const EdgeInsets.only(left: CtSpacing.m),
-      child: Text(
-        l10n.production_labourQueued(data.queuedCount),
-        style: theme.textTheme.labelSmall,
-      ),
-    );
-  }
-
-  List<Widget> _buildEditActions(String tierName) {
-    final disbandLabel = l10n.production_labourDisband;
-    return [
-      _LabourIconButton(
-        enabled: data.canPop,
-        semanticLabel: l10n.production_labourDequeueTier(tierName),
-        tooltip: l10n.production_labourDequeueTier(tierName),
-        assetFileName: _uiIconLabourDecrement,
-        onPressed: () => callbacks.onPopLastRecruitOrder(data.tier),
-      ),
-      _LabourIconButton(
-        enabled: data.canAppend,
-        semanticLabel: _appendTooltip(),
-        tooltip: _appendTooltip(),
-        assetFileName: _uiIconLabourIncrement,
-        onPressed: () => callbacks.onAppendRecruitOrder(data.tier),
-      ),
-      if (data.tier != WorkerTier.peasant)
-        _DisbandTierButton(
-          tier: data.tier,
-          enabled: data.canDisband,
-          disbandLabel: disbandLabel,
-          tooltip: l10n.production_labourDisbandTier(tierName),
-          onDisband: callbacks.onDisband,
-        )
-      else
-        // Peasant rows have no visible Disband control but reserve the
-        // same trailing slot width as trained rows so −/+ align across
-        // all tier rows. SPEC/ui/production-panel.md § Labour Controls
-        // (12-A) > Trailing alignment. Refs #2862 S8a / C4 / G5.
-        _DisbandReservedSlot(label: disbandLabel),
-    ];
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final tierName = _tierName();
-    final tierLabel = _tierLabelWithUnlockState();
-    // Expanded (not Flexible) on the leading tier-label slot so the slot
-    // width is the same across rows. With Flexible.loose the label slot
-    // collapses to the text's intrinsic width, which differs per tier
-    // (e.g. "Peasants (unlocked)" vs "Apprentices (unlocked)") and
-    // shifts the trailing action cluster horizontally — breaking the
-    // trailing-alignment contract in SPEC/ui/production-panel.md
-    // § Labour Controls (12-A) > Trailing alignment. With Expanded, the
-    // label slot occupies all available pre-cluster space, the cluster
-    // anchors flush to the row's right edge, and combined with the
-    // reserved-Disband-slot on the peasant row the −/+ steppers sit at
-    // the same screen-x across every tier row. Refs #2862 S8a / C4 / G5.
-    return Row(
-      mainAxisSize: MainAxisSize.max,
-      children: [
-        Expanded(
-          child: Row(
-            mainAxisSize: MainAxisSize.max,
-            children: [
-              Flexible(
-                child: Text(
-                  tierLabel,
-                  style: theme.textTheme.bodySmall,
-                  overflow: TextOverflow.ellipsis,
-                ),
-              ),
-              _buildQueuedSegment(),
-            ],
-          ),
-        ),
-        if (canEdit) ..._buildEditActions(tierName),
-      ],
-    );
-  }
-}
-
-class _DisbandTierButton extends StatelessWidget {
-  const _DisbandTierButton({
-    required this.tier,
-    required this.enabled,
-    required this.disbandLabel,
-    required this.tooltip,
-    required this.onDisband,
-  });
-
-  final WorkerTier tier;
-  final bool enabled;
-  final String disbandLabel;
-  final String tooltip;
-  final void Function(WorkerTier tier) onDisband;
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.only(left: CtSpacing.s),
-      child: CtDangerTextButton(
-        key: ValueKey<String>('production_labour_disband_${tier.id}'),
-        enabled: enabled,
-        label: disbandLabel,
-        semanticLabel: tooltip,
-        tooltip: tooltip,
-        onPressed: enabled ? () => onDisband(tier) : null,
-      ),
-    );
-  }
-}
-
-/// Invisible Disband-shaped placeholder used by the peasant row to reserve
-/// the same trailing slot width as trained rows so −/+ step buttons sit at
-/// the same screen-x coordinates across every Labour row.
-///
-/// SPEC/ui/production-panel.md § Labour Controls (12-A) > Trailing
-/// alignment (Refs #2862 S8a / C4 / G5). The reserved slot:
-///
-/// - Carries no widget key (so key-based finders for
-///   `production_labour_disband_<tier>` skip it).
-/// - Wraps the *enabled* [CtDangerTextButton] (with a no-op callback) in
-///   [Opacity] 0.0 so the slot matches the exact rendered dimensions of a
-///   trained-tier Disband button (including the Material / InkWell padded
-///   tap-target inset on enabled buttons) but paints no visible pixels.
-///   A disabled `CtDangerTextButton` would skip the Material / InkWell
-///   wrapping and produce a narrower intrinsic width than the real
-///   enabled Disband — that would mis-align the peasant row's −/+
-///   steppers.
-/// - Wraps the subtree in [ExcludeSemantics] + [IgnorePointer] so screen
-///   readers do not announce a second "Disband" affordance on peasant
-///   rows and pointer events are ignored.
-class _DisbandReservedSlot extends StatelessWidget {
-  const _DisbandReservedSlot({required this.label});
-
-  final String label;
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.only(left: CtSpacing.s),
-      child: ExcludeSemantics(
-        child: IgnorePointer(
-          child: Opacity(
-            opacity: 0.0,
-            child: CtDangerTextButton(
-              enabled: true,
-              label: label,
-              onPressed: _noop,
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-
-  static void _noop() {}
-}
-
-/// Per-tier `+` / `−` control for the Labour Controls section.
-///
-/// Renders the shared dark editorial-monocle 26 × 26 step-button surface
-/// ([ProductionStepButtonSurface]) so the Available subpanel's per-tier
-/// recruit/train controls reuse the same chrome as the Allocation
-/// subpanel's per-recipe ± / maximize / clear controls — `SPEC/ui/production-panel.md`
-/// § Allocation step buttons explicitly mandates this contract reuse
-/// (`Refs #2862` § Labour Controls).
-///
-/// The surface fades to [kProductionAllocationStepButtonDisabledOpacity]
-/// when [enabled] is false; tap gestures are gated by the same flag so
-/// disabled controls never dispatch [onPressed].
-class _LabourIconButton extends StatelessWidget {
-  const _LabourIconButton({
-    required this.enabled,
-    required this.semanticLabel,
-    required this.tooltip,
-    required this.assetFileName,
-    required this.onPressed,
-  });
-
-  static const double _iconSize = 15;
-
-  final bool enabled;
-  final String semanticLabel;
-  final String tooltip;
-  final String assetFileName;
-  final VoidCallback onPressed;
-
-  @override
-  Widget build(BuildContext context) {
-    final path = '$kAppIconAssetPrefix$assetFileName';
-    final surface = ProductionStepButtonSurface(
-      enabled: enabled,
-      iconAssetPath: path,
-      iconSize: _iconSize,
-    );
-    return Semantics(
-      button: true,
-      enabled: enabled,
-      label: semanticLabel,
-      child: Tooltip(
-        message: tooltip,
-        child: Material(
-          color: Colors.transparent,
-          child: InkWell(onTap: enabled ? onPressed : null, child: surface),
-        ),
-      ),
     );
   }
 }
